client-flask/requirements.txt
flask==3.0.0
stomp.py==8.1.0
prometheus-client==0.19.0
python-dateutil==2.8.2
gunicorn==21.2.0
Werkzeug==3.0.1
